Sync

+++ src/deserialize/cache.rs
@@ -34,6 +34,6 @@ impl Drop for CachedKey {
 }
 
 pub type KeyMap =
-    AssociativeCache<u32, CachedKey, Capacity512, HashDirectMapped, RoundRobinReplacement>;
+    AssociativeCache<u64, CachedKey, Capacity512, HashDirectMapped, RoundRobinReplacement>;
 
 pub static mut KEY_MAP: OnceCell<KeyMap> = OnceCell::new();
